Updating Checkout Actions Task (#758)
This pull request makes a minor update to the GitHub Actions workflow by upgrading the `actions/checkout` action from version 4 to version 5 in the `.github/workflows/build.yml` file. This ensures that the workflow uses the latest features and bug fixes from the action. It ensures consistency across the different pipelines.